+++ CHANGELOG.md
@@ -31,12 +31,14 @@ Test/interim releases are not documented here.
 
 ## WIP
 ### Add
-- Add convenience helpers for `SmTransformer`.
-- Add diagram based toml config for `RenderConfig` and *some* `SmRunner.Settings`
+- Add diagram based toml config for `RenderConfig` and `SmRunner.Settings`
   - Useful for plantuml as well which previously never had diagram based render config support.
   - https://github.com/StateSmith/StateSmith/issues/335
 - simulator - show implicit `do` edge trigger explicitly
     - https://github.com/StateSmith/StateSmith/issues/316
+- Add default expansions for variables and functions.
+    - https://github.com/StateSmith/StateSmith/issues/284
+- Add convenience helpers for `SmTransformer`.
 
 # Fix
 - simulator - prevent user diagram settings that could mess up generated simulation.